Extract profileEdit.ts profile/segment helpers from projectStore (+ tests)
